Prevent routes file write if content is identical
This keeps webpack-dev-server from recompiling when the content of the
generated routes file didn't change. I looked into using
`app.routes_reloader` to detect changes, but that didn't work when the
routes file(s) reference other classes or modules (which I think is a
somewhat common way of splitting apart large routes files).

I wanted to keep using `File.atomic_write`, so this uses throw/catch to
prevent the write if the tempfile is identical to the existing file.

Addresses the same issue as: https://github.com/davishmcclurg/webpacker-routes/pull/8